Rework for cpp-sensor 0.7.0 and nginx 1.17.6
The files from the nginx-opentracing project cannot be used any
more as they could become incompatible. So use a plain Ubuntu 18.04
Bionic image instead. Set up the official Nginx mainline repository
and install Nginx 1.17.6. Adapt the nginx image Dockerfile and the
Nginx config to run with output to stdout. Download the Instana
ngx_http_opentracing_module for Nginx 1.17.6 as well. Adapt the
service name to "nginxtracing_nginx" to unify it with the Docker
image name.