binaries: Fix issues found during review
Some spelling and punctuation is not correct. Then the NGINX Plus
links do not work. There are still links to .so files. The structure
is not correct as the compatnfo links belong to the NGINX official
repository as well. So address those issues.

V2: Fix "libinstana_sensor" to "nginx_tracing" in Artifactory paths.