Release 1.0.0 (#22)
* README: Update release history for 1.0.0

* README: Drop technology preview remarks

This is GA now. So drop technology preview remarks.

* nginx/Dockerfile: Simplify updating of the NGINX version

There are two locations where the NGINX version is used. So put it
into an environment variable. Also be verbose about the sensor
version used.

* nginx/nginx.conf: Adapt to new documentation

Drop the NGINX OpenTracing configuration directives which are the
new default already and move "opentracing_propagate_context" to
the http top level.

* nginx/Dockerfile: Download the zip bundle

If customers download the .so files separately, then this has the
chance of interoperability issues. So show the .zip bundle method
instead.

* README: Address issues found earlier in Instana docs

The text for the Instana docs has been copied from here.
So just fix spelling, word order, and missing or superfluous words.

* README: Consistently spell NGINX uppercase

* Extract libc version to an env var

* Dockerfile as multi-step build; support download key

* Interpolate empty downloadkey arg to empty string

Co-authored-by: Michele Mancioppi <michele.mancioppi@gmail.com>