docker-compose: Pull the Instana agent from icr.io
Due to IBM policies, the Instana agent image may not be uploaded to
DockerHub anymore. The image was still cached on my machine though.
But for other users `docker-compose up` results in an error when
pulling the image. The Instana agent images are uploaded to icr.io
now.

So update the agent image config with the information from the
Instana UI "Installing Instana Agents" -> "Docker".

Reported-by: Yang Kwang Goh <gohykw@sg.ibm.com>